filter-apply-form should now work without data-power-select.
