--- app/components/chart-ranking-single.hbs
@@ -1,45 +1,0 @@
-<div>
-  <Input
-    @type="checkbox" @checked={{this.showAllFilterGroups}}
-    name="showAllFilterGroups" id="showAllFilterGroupsCheckbox" />
-  <label for="showAllFilterGroupsCheckbox">Show all filter groups</label>
-</div>
-
-<FilterApplyForm
-  @filterGroups={{@filterGroups}}
-  @appliedFilterObjs={{@appliedFilterObjs}}
-  @appliedFiltersString={{@appliedFiltersString}}
-  @updateAppliedFiltersString={{@updateAppliedFiltersString}} />
-
-<table class="records-table">
-  <thead>
-    <tr>
-      <th>Rank</th>
-      <th>Player</th>
-      <th>Record</th>
-      {{#each this.shownFilterGroups as |filterGroup|}}
-        <th>{{filterGroup.name}}</th>
-      {{/each}}
-    </tr>
-  </thead>
-
-  <tbody>
-    {{#each @records as |record|}}
-      <tr>
-        <td>{{record.rank}}</td>
-        <td>{{record.player_username}}</td>
-        <td>
-          <LinkTo @route="charts.player-history" @models={{array @chart.id record.player_id}}>
-            {{record.value_display}}
-          </LinkTo>
-        </td>
-
-        {{#each this.shownFilterGroups as |filterGroup|}}
-          <td>
-            {{get (find-filter-of-group record.filters filterGroup.id) "name"}}
-          </td>
-        {{/each}}
-      </tr>
-    {{/each}}
-  </tbody>
-</table>
