+++ app/controllers/charts/show.js
@@ -1,4 +1,8 @@
+import { A } from '@ember/array';
 import Controller from '@ember/controller';
+import { action } from '@ember/object';
+import { tracked } from '@glimmer/tracking';
+
 
 export default class ChartsShowController extends Controller {
   queryParams = [{
@@ -7,4 +11,34 @@ export default class ChartsShowController extends Controller {
     appliedFiltersString: 'filters'
   }];
   appliedFiltersString = null;
+
+  @tracked showAllFilterGroups = null;
+  @tracked selectedFilterGroup = null;
+  @tracked filterOptions = A([]);
+
+  get shownFilterGroups() {
+    if (this.showAllFilterGroups) {
+      return this.model.filterGroups;
+    }
+    else {
+      return this.model.filterGroups.filterBy('showByDefault', true);
+    }
+  }
+
+  @action
+  updateSelectedFilterGroup(newSelectedFilterGroup) {
+    this.selectedFilterGroup = newSelectedFilterGroup;
+
+    this.filterOptions = this.store.query('filter', {
+      filter_group_id: newSelectedFilterGroup.id,
+    })
+  }
+
+  @action
+  onSearchTextChange(searchText) {
+    this.filterOptions = this.store.query('filter', {
+      filter_group_id: this.selectedFilterGroup.id,
+      name_search: searchText,
+    })
+  }
 }
