+++ app/templates/charts/show.hbs
@@ -21,9 +21,52 @@
 </div>
 
 
-<ChartRankingSingle
-  @chart={{@model.chart}} @records={{@model.records}}
+<div>
+  <Input
+    @type="checkbox" @checked={{this.showAllFilterGroups}}
+    name="showAllFilterGroups" id="showAllFilterGroupsCheckbox" />
+  <label for="showAllFilterGroupsCheckbox">Show all filter groups</label>
+</div>
+
+<FilterApplyForm
   @filterGroups={{@model.filterGroups}}
   @appliedFilterObjs={{@model.appliedFilterObjs}}
-  @appliedFiltersString={{appliedFiltersString}}
-  @updateAppliedFiltersString={{action (mut appliedFiltersString)}} />
+  @appliedFiltersString={{this.appliedFiltersString}}
+  @updateAppliedFiltersString={{action (mut this.appliedFiltersString)}}
+  @selectedFilterGroup={{this.selectedFilterGroup}}
+  @controllerUpdateSelectedFilterGroup={{action this.updateSelectedFilterGroup}}
+  @filterOptions={{this.filterOptions}}
+  @onSearchTextChange={{action this.onSearchTextChange}} />
+
+<table class="records-table">
+  <thead>
+    <tr>
+      <th>Rank</th>
+      <th>Player</th>
+      <th>Record</th>
+      {{#each this.shownFilterGroups as |filterGroup|}}
+        <th>{{filterGroup.name}}</th>
+      {{/each}}
+    </tr>
+  </thead>
+
+  <tbody>
+    {{#each @model.records as |record|}}
+      <tr>
+        <td>{{record.rank}}</td>
+        <td>{{record.player_username}}</td>
+        <td>
+          <LinkTo @route="charts.player-history" @models={{array @model.chart.id record.player_id}}>
+            {{record.value_display}}
+          </LinkTo>
+        </td>
+
+        {{#each this.shownFilterGroups as |filterGroup|}}
+          <td>
+            {{get (find-filter-of-group record.filters filterGroup.id) "name"}}
+          </td>
+        {{/each}}
+      </tr>
+    {{/each}}
+  </tbody>
+</table>
